Refactor test_groupby and test_query into their own files
