+++ thicket/tests/test_filter.py
@@ -4,9 +4,7 @@
 # SPDX-License-Identifier: MIT
 
 import pytest
-import re
 
-import hatchet as ht
 from thicket import Thicket, InvalidFilter, EmptyMetadataFrame
 
 
@@ -135,113 +133,3 @@ def test_filter_stats(example_cali):
             )
             # check filtered ensembleframe nodes match exp_nodes
             assert ensemble_nodes == exp_nodes
-
-
-def test_groupby(example_cali):
-    # example thicket
-    th = Thicket.from_caliperreader(example_cali)
-    # use cases for string, numeric, and single value columns
-    columns = ["user", "launchdate", "cali.channel"]
-
-    # inspect all use cases
-    for col_value in columns:
-        unique_values = sorted(th.metadata[col_value].unique().tolist())
-        th_list = th.groupby(col_value)
-        # inspect all unique values in the use case
-        for itr, uni_val in enumerate(unique_values):
-            # expected profile hashes and nodes
-            exp_profiles = sorted(
-                th.metadata.index[th.metadata[col_value] == uni_val].tolist()
-            )
-            exp_nodes = sorted(
-                th.dataframe[th.dataframe.index.get_level_values(1).isin(exp_profiles)]
-                .index.get_level_values(0)
-                .drop_duplicates()
-                .tolist()
-            )
-
-            # metadata profile hash for unique value sub-thicket
-            metadata_profiles = sorted(th_list[itr].metadata.index.tolist())
-            assert metadata_profiles == exp_profiles
-
-            # ensemble profile hash for unique value sub-thicket
-            ensemble_profiles = sorted(
-                th_list[itr]
-                .dataframe.index.get_level_values(1)
-                .drop_duplicates()
-                .tolist()
-            )
-            assert ensemble_profiles == exp_profiles
-
-            # ensemble nodes for unique value sub-thicket
-            ensemble_nodes = sorted(
-                th_list[itr]
-                .dataframe.index.get_level_values(0)
-                .drop_duplicates()
-                .tolist()
-            )
-            assert ensemble_nodes == exp_nodes
-
-            # stats nodes for unique value sub-thicket
-            stats_nodes = sorted(
-                th_list[itr]
-                .statsframe.dataframe.index.get_level_values(0)
-                .drop_duplicates()
-                .tolist()
-            )
-            assert stats_nodes == exp_nodes
-
-            # check for name column statsframe dataframe
-            assert "name" in th_list[itr].statsframe.dataframe.columns
-
-    # drop all rows of the metadataframe
-    th.metadata = th.metadata.iloc[0:0]
-
-    # check for empty metadataframe exception
-    with pytest.raises(EmptyMetadataFrame):
-        th.groupby(["user"])
-
-
-def test_query(rajaperf_basecuda_xl_cali):
-    th = Thicket.from_caliperreader(rajaperf_basecuda_xl_cali)
-    th_df_profiles = th.dataframe.index.get_level_values("profile")
-    match = [
-        node
-        for node in th.graph.traverse()
-        if node._hatchet_nid in [0, 1, 2, 3, 5, 6, 8, 9]
-    ]
-    match_frames = [node.frame for node in match]
-    match_names = [frame["name"] for frame in match_frames]
-    query = (
-        ht.QueryMatcher()
-        .match("*")
-        .rel(
-            ".",
-            lambda row: row["name"]
-            .apply(lambda x: re.match(r"Algorithm.*block_128", x) is not None)
-            .all(),
-        )
-    )
-    filt_th = th.query(query)
-    filt_nodes = list(filt_th.graph.traverse())
-
-    assert sorted(list(filt_th.dataframe.index.names)) == sorted(["node", "profile"])
-    filt_th_df_nodes = filt_th.dataframe.index.get_level_values("node").to_list()
-    filt_th_df_profiles = filt_th.dataframe.index.get_level_values("profile")
-
-    assert len(filt_nodes) == len(match)
-    assert all([n.frame in match_frames for n in filt_nodes])
-    assert all([n.frame["name"] in match_names for n in filt_nodes])
-    assert all([n in filt_th_df_nodes for n in filt_nodes])
-    assert sorted(filt_th_df_profiles.unique().to_list()) == sorted(
-        th_df_profiles.unique().to_list()
-    )
-
-    # update_inclusive_columns is called automatically by GraphFrame.squash()
-    # at the end of "query".
-    # To make sure the following assertion works, manually invoke this function
-    # on the original data.
-    # TODO if/when the call to update_inclusive_columns is made optional in Hatchet,
-    #      remove this line
-    th.update_inclusive_columns()
-    assert sorted(list(filt_th.dataframe.columns)) == sorted(list(th.dataframe.columns))
