+++ thicket/tests/test_groupby.py
@@ -1,0 +1,73 @@
+# Copyright 2022 Lawrence Livermore National Security, LLC and other
+# Thicket Project Developers. See the top-level LICENSE file for details.
+#
+# SPDX-License-Identifier: MIT
+
+import pytest
+
+from thicket import Thicket, EmptyMetadataFrame
+
+
+def test_groupby(example_cali):
+    # example thicket
+    th = Thicket.from_caliperreader(example_cali)
+    # use cases for string, numeric, and single value columns
+    columns = ["user", "launchdate", "cali.channel"]
+
+    # inspect all use cases
+    for col_value in columns:
+        unique_values = sorted(th.metadata[col_value].unique().tolist())
+        th_list = th.groupby(col_value)
+        # inspect all unique values in the use case
+        for itr, uni_val in enumerate(unique_values):
+            # expected profile hashes and nodes
+            exp_profiles = sorted(
+                th.metadata.index[th.metadata[col_value] == uni_val].tolist()
+            )
+            exp_nodes = sorted(
+                th.dataframe[th.dataframe.index.get_level_values(1).isin(exp_profiles)]
+                .index.get_level_values(0)
+                .drop_duplicates()
+                .tolist()
+            )
+
+            # metadata profile hash for unique value sub-thicket
+            metadata_profiles = sorted(th_list[itr].metadata.index.tolist())
+            assert metadata_profiles == exp_profiles
+
+            # ensemble profile hash for unique value sub-thicket
+            ensemble_profiles = sorted(
+                th_list[itr]
+                .dataframe.index.get_level_values(1)
+                .drop_duplicates()
+                .tolist()
+            )
+            assert ensemble_profiles == exp_profiles
+
+            # ensemble nodes for unique value sub-thicket
+            ensemble_nodes = sorted(
+                th_list[itr]
+                .dataframe.index.get_level_values(0)
+                .drop_duplicates()
+                .tolist()
+            )
+            assert ensemble_nodes == exp_nodes
+
+            # stats nodes for unique value sub-thicket
+            stats_nodes = sorted(
+                th_list[itr]
+                .statsframe.dataframe.index.get_level_values(0)
+                .drop_duplicates()
+                .tolist()
+            )
+            assert stats_nodes == exp_nodes
+
+            # check for name column statsframe dataframe
+            assert "name" in th_list[itr].statsframe.dataframe.columns
+
+    # drop all rows of the metadataframe
+    th.metadata = th.metadata.iloc[0:0]
+
+    # check for empty metadataframe exception
+    with pytest.raises(EmptyMetadataFrame):
+        th.groupby(["user"])
